backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pypdf
tiktoken
python-multipart
numpy
sentence-transformers
langchain-huggingface
tf-keras
langchain-google-genai
sqlalchemy
psycopg2-binary
